Feat | New UI

+++ app/src/main/java/com/example/blemeter/core/ble/domain/model/request/PurchaseDataRequest.kt
@@ -1,6 +1,6 @@
 package com.example.blemeter.core.ble.domain.model.request
 
 data class PurchaseDataRequest(
-    val numberTimes: UInt = 0u,
-    val purchaseVariable: UInt = 0u
+    val numberTimes: Int = 0,
+    val purchaseVariable: Double = 0.0
 )
